Return images for characters
The thing is that FastAPI is not created to
work under proxy. To implment everything properly
we need to pass X-Forwarded-Proto header.

Headers is another task, because it depends on
proxy. If you are going to implement this part,
please check Django source code, you will find
how to make everything in the best way.